requirements.txt
requests
json2html
yara-python
colorama
jinja2
pyscreeze